Refactored sub-classes (clothing, decor, electronics) to inherit from Item parent
